+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
+- Change TypeScript-example to match other examples. 
 ### Deprecated
 ### Removed
 ### Fixed